Add files via upload
This class converts any html webpage to a .txt file which can be easily embedded as tokens in train/test data-set of a Transformer-based Language Model. While this model is almost foolproof for natural language modules, there are still a few visible bugs within the space of mathematical symbols and semantics module. Fixing the latter is the goal of a future debug iteration; for now, we can approximate that in the limit of training a Transformer-based Language Model on the space of thousands, no, millions of such research papers, a fine-tuned model would wash out the influences of minor symbolic bugs in each paper, as the test/train data set in such an ideal case would naturally incur some visible covariance (effectively, standard error around the mean) which would absorb these bugs and uncertainties within the multi-dimensional token space. Having said that, such bugs may as well lead to hallucinations, and it is up to us to determine how these effects propagate through the model upon training.